move generating unique fields in DB logic from `TesterAccountService` into separate `UniqueFieldGenerator` class
